Remove sawyer_env.py
- Remove it because we have abstracted the robot interface.
