+++ contrib/ros/envs/sawyer/example_launchers/trpo_sim_sawyer_push.py
@@ -11,22 +11,22 @@ from sandbox.rocky.tf.algos.trpo import TRPO
 from sandbox.rocky.tf.policies.gaussian_mlp_policy import GaussianMLPPolicy
 from sandbox.rocky.tf.envs.base import TfEnv
 
-from contrib.ros.envs.example_launchers import model_dir
-from contrib.ros.envs.sawyer.pick_and_place_env import PickAndPlaceEnv
+from contrib.ros.envs.sawyer.example_launchers import model_dir
+from contrib.ros.envs.sawyer.push_env import PushEnv
 from contrib.ros.util.task_object_manager import TaskObject, TaskObjectManager
 
 
 def run_task(*_):
-    block = TaskObject(
-        name='block',
-        initial_pos=(0.5725, 0.1265, 0.90),
-        random_delta_range=0.15,
-        resource=osp.join(model_dir, 'block/model.urdf'))
     table = TaskObject(
         name='table',
         initial_pos=(0.75, 0.0, 0.0),
         random_delta_range=0.15,
         resource=osp.join(model_dir, 'cafe_table/model.sdf'))
+    block = TaskObject(
+        name='block',
+        initial_pos=(0.5725, 0.1265, 0.90),
+        random_delta_range=0.15,
+        resource=osp.join(model_dir, 'block/model.urdf'))
 
     initial_goal = np.array([0.6, -0.1, 0.80])
 
@@ -43,13 +43,13 @@ def run_task(*_):
 
     rospy.init_node('trpo_sim_sawyer_pnp_exp', anonymous=True)
 
-    pnp_env = PickAndPlaceEnv(initial_goal, task_obj_mgr, simulated=True)
+    push_env = PushEnv(initial_goal, task_obj_mgr, simulated=True)
 
-    rospy.on_shutdown(pnp_env.shutdown)
+    rospy.on_shutdown(push_env.shutdown)
 
-    pnp_env.initialize()
+    push_env.initialize()
 
-    env = TfEnv(normalize(pnp_env))
+    env = TfEnv(normalize(push_env))
 
     policy = GaussianMLPPolicy(
         name="policy", env_spec=env.spec, hidden_sizes=(32, 32))
