+++ contrib/ros/envs/sawyer/pick_and_place_env.py
@@ -6,47 +6,30 @@ import numpy as np
 
 from rllab.core.serializable import Serializable
 
-from contrib.ros.envs import sawyer_env
+from contrib.ros.envs.ros_env import RosEnv
 from contrib.ros.robots.sawyer import Sawyer
 
-INITIAL_ROBOT_JOINT_POS = {
-    'right_j0': -0.041662954890248294,
-    'right_j1': -1.0258291091425074,
-    'right_j2': 0.0293680414401436,
-    'right_j3': 2.17518162913313,
-    'right_j4': -0.06703022873354225,
-    'right_j5': 0.3968371433926965,
-    'right_j6': 1.7659649178699421,
-}
 
-
-class PickAndPlaceEnv(sawyer_env.SawyerEnv, Serializable):
+class PickAndPlaceEnv(RosEnv, Serializable):
     def __init__(self,
                  initial_goal,
                  task_obj_mgr,
                  sparse_reward=True,
                  simulated=False,
                  distance_threshold=0.05,
-                 target_range=0.15,
-                 target_in_the_air=False):
+                 target_range=0.15):
         Serializable.quick_init(self, locals())
 
         self._distance_threshold = distance_threshold
         self._target_range = target_range
         self._sparse_reward = sparse_reward
-        self._target_in_the_air = target_in_the_air
         self.initial_goal = initial_goal
         self.goal = self.initial_goal.copy()
 
         sawyer = Sawyer(simulated=simulated, control_mode='position')
 
-        sawyer_env.SawyerEnv.__init__(
-            self,
-            task_obj_mgr=task_obj_mgr,
-            robot=sawyer,
-            has_object=True,
-            simulated=True,
-            obj_range=0.15)
+        RosEnv.__init__(
+            self, task_obj_mgr=task_obj_mgr, robot=sawyer, simulated=simulated)
 
     def sample_goal(self):
         """
